Revisions and documentation
What I did:

1) Although the function worked correctly, there were parts of the
function that were causing warning messages (like "number of items to
replace is not a multiple of replacement length") because of
“technically” incorrect coding. I corrected that.

2) The function works when NULL is assigned to any of all.nothing,
either.or, and always.

3) I revised the function to work under parallel when specified.

4) I tried many things in many ways, but could not change the for loop
to a plyr function. I think it’s okay to leave the for loop as it’s a
simple code.

5) Originally, I could not get the function to work when all variables
are conditioned. I got the following error: Error in do.ply(i) :  task
2 failed - "attempt to select less than one element". However, I solved
the problem by making the function create the “truefalse.list” only
when the length of “therest” is not zero. Now the function works even
when all variables are conditioned.