modelSelect function moved into fitBMA.R
The modelSelect function works fine. But one very important problem is
that the interaction term actually replaces an existing variable rather
than being a separate term. This happens because in order to calculate
the model configurations for an interaction term, the user specifies an
existing variable as the interaction term. For instance, say the user
has 10 variables. The variables specified in the “interactions” object
is X1, X2, and X3. This means that the user wants X3 to be an
interaction of X1 and X2. However, the code works so that X3 is the X3
variable the user has!! This can probably be solved by changing the
code so that the matrix of model configurations expands by a single
column to incorporate the newly created interaction term.