Scripts for presentation
I’ve included the scripts to be used to explain the modelSelect
function. The file named “PreRunModelSelectScript” is the one that
should be run before the presentation to save the objects. The file
named “InClassModelSelectScript” is the one to use during class to
demonstrate the flow of the function and its key parts.